+++ src/main/java/com/example/Security/Repository/TokenRepository.java
@@ -7,7 +7,7 @@ import org.springframework.data.jpa.repository.Query;
 import java.util.List;
 import java.util.Optional;
 
-public interface TokenRepository extends JpaRepository<Token, Integer> {
+public interface TokenRepository extends JpaRepository<Token, Long> {
 
 
     @Query("""
@@ -15,7 +15,7 @@ public interface TokenRepository extends JpaRepository<Token, Integer> {
             select t from Token t inner join User u on t.user.id = u.id
 where t.user.id = :userId and t.loggedOut = false
 """)
-    List<Token> findAllAccessTokensByUser(Integer userId);
+    List<Token> findAllAccessTokensByUser(Long userId);
 
     Optional<Token> findByToken(String token);
